seperate types, styled component and constants from main Button file
