+++ uikit/Button/index.tsx
@@ -18,83 +18,11 @@
  */
 
 import * as React from 'react';
-import PropTypes, { any } from 'prop-types';
-import styled from '@emotion/styled';
-import { css } from '@emotion/core';
 import Icon from '../Icon';
 import useTheme from '../utils/useTheme';
-import FocusWrapper from '../FocusWrapper';
-
-type ButtonVariant = 'primary' | 'secondary' | 'text';
-export type ButtonSize = 'sm' | 'md';
-
-export const BUTTON_VARIANTS: {
-  PRIMARY: ButtonVariant;
-  SECONDARY: ButtonVariant;
-  TEXT: ButtonVariant;
-} = Object.freeze({
-  PRIMARY: 'primary',
-  SECONDARY: 'secondary',
-  TEXT: 'text',
-});
-
-export const BUTTON_SIZES: {
-  SM: ButtonSize;
-  MD: ButtonSize;
-} = Object.freeze({
-  SM: 'sm',
-  MD: 'md',
-});
-
-const StyledButton = styled<
-  typeof FocusWrapper,
-  {
-    size: 'sm' | 'md';
-    variant: 'primary' | 'secondary' | 'text';
-    disabled: boolean;
-  }
->(FocusWrapper)`
-  ${({ theme }) => css(theme.typography.default)};
-  transition: all 0.25s;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  outline: none;
-  box-shadow: none;
-  border: none;
-  border-radius: 100px;
-  font-weight: 600;
-  text-transform: uppercase;
-  border-style: solid;
-
-  border-width: ${({ theme, size }) => theme.button.borderWeights[size]};
-  padding: ${({ theme, size }) => theme.button.paddings[size]};
-  color: ${({ theme, variant }) => theme.button.textColors[variant].default};
-  cursor: ${({ disabled }) => (disabled ? 'not-allowed' : 'pointer')};
-  font-size: ${({ theme, size }) => theme.button.fontSizes[size]};
-  background-color: ${({ theme, variant }) => theme.button.colors[variant].default};
-  border-color: ${({ theme, variant }) => theme.button.borderColors[variant].default};
-
-  &:focus {
-    background-color: ${({ theme, variant }) => theme.button.colors[variant].focus};
-  }
-
-  &:hover {
-    background-color: ${({ theme, variant }) => theme.button.colors[variant].hover};
-    border-color: ${({ theme, variant }) => theme.button.borderColors[variant].hover};
-  }
-
-  &:active {
-    background-color: ${({ theme, variant }) => theme.button.colors[variant].active};
-    border-color: ${({ theme, variant }) => theme.button.borderColors[variant].active};
-  }
-
-  &:disabled {
-    background-color: ${({ theme, variant }) => theme.button.colors[variant].disabled};
-    border-color: ${({ theme, variant }) => theme.button.borderColors[variant].disabled};
-    color: ${({ theme, variant }) => theme.button.textColors[variant].disabled};
-  }
-`;
+import { BUTTON_VARIANTS, BUTTON_SIZES } from './constants';
+import StyledButton from './styled';
+import { ButtonVariant, ButtonSize } from './types';
 
 const Button = React.forwardRef<
   HTMLButtonElement,
