+++ uikit/Button/types.ts
@@ -1,0 +1,7 @@
+export type ButtonVariant = 'primary' | 'secondary' | 'text';
+export type ButtonSize = 'sm' | 'md';
+export type ButtonLoader = {
+  shouldShowLoading: boolean;
+  variant: string;
+  theme: any;
+};
